news-framing-project/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cssselect==1.4.0
exceptiongroup==1.3.1
fastapi==0.136.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.29.0
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.13
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.3
lxml==6.1.0
lxml_html_clean==0.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.4
numpy==2.2.6
packaging==26.1
pandas==2.3.3
pillow==12.2.0
psycopg2-binary==2.9.12
py4j==0.10.9.9
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyspark==4.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-file==3.0.1
rich==15.0.0
safetensors==0.7.0
schedule==1.2.2
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
starlette==1.0.0
sympy==1.14.0
textblob==0.20.0
tinysegmenter==0.3
tldextract==5.3.1
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.6.2
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uvicorn==0.46.0
